+++ source/core/audio/windows/audio_sys_win_base.cpp
@@ -14,8 +14,8 @@ HRESULT GetBufferAndCallBackClient(
     int & audio_result)
 {
 
-    size_t space_remaining = 0;
-    size_t adding = 0;
+    uint32_t space_remaining = 0;
+    uint32_t adding = 0;
     HRESULT  hr = S_OK;
     BYTE *pData;
     DWORD flags = 0;
@@ -42,7 +42,7 @@ HRESULT GetBufferAndCallBackClient(
 
     audioBuff.currentSize += adding;
 
-    // If we there is data remaining in the audio buffer pData.
+    // If there is data remaining in the audio buffer pData:
     // Flush the local audio buff and reset it. Then copy over the remaining data.
     if (adding < (numFramesAvailable * 2))
     {
@@ -50,13 +50,13 @@ HRESULT GetBufferAndCallBackClient(
             audioBuff.pAudioData,
             audioBuff.currentSize);
 
-        size_t data_remaining = (numFramesAvailable * 2) - adding;
+        uint32_t data_remaining = (numFramesAvailable * 2) - adding;
 
         // Reset local audio buffer.
         audioBuff.currentSize = 0;
         if (data_remaining > 0)
         {
-            size_t copySize = min(data_remaining, audioBuff.totalSize);
+            uint32_t copySize = min(data_remaining, audioBuff.totalSize);
             memcpy_s(audioBuff.pAudioData,
                 audioBuff.totalSize,
                 pData + adding,
